+++ src/sum-exercise.js
